requirements.txt
amqp==2.6.0
APScheduler==3.6.3
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-mysql==3.5.0
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
future==0.18.2
idna==2.9
importlib-metadata==1.6.1
kombu==4.6.10
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
uwsgi
